backend/requirements.txt
fastapi
uvicorn[standard]
websockets
sqlmodel
python-dotenv
shortuuid